# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "학원 출결 관리 시스템을 위한 RESTful API",
        "title": "출결 관리 시스템 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/attendance/{date}": {
            "get": {
                "description": "특정 날짜의 모든 학생 출결 정보를 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attendance"
                ],
                "summary": "날짜별 전체 출결 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "날짜 (YYYY-MM-DD)",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Attendance"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/students/{student_id}/attendance/{date}": {
            "get": {
                "description": "특정 학생의 특정 날짜 출결 정보를 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attendance"
                ],
                "summary": "학생별 특정 날짜 출결 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "학생 ID",
                        "name": "student_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "날짜 (YYYY-MM-DD)",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Attendance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Attendance not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "특정 학생의 특정 날짜 출결을 수정합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attendance"
                ],
                "summary": "출결 수정",
                "parameters": [
                    {
                        "type": "string",
                        "description": "학생 ID",
                        "name": "student_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "날짜 (YYYY-MM-DD)",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "수정할 출결 정보",
                        "name": "attendance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Attendance"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Attendance not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "특정 학생의 특정 날짜 출결을 등록합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attendance"
                ],
                "summary": "출결 등록",
                "parameters": [
                    {
                        "type": "string",
                        "description": "학생 ID",
                        "name": "student_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "날짜 (YYYY-MM-DD)",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "출결 정보",
                        "name": "attendance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Attendance"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "특정 학생의 특정 날짜 출결을 삭제합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attendance"
                ],
                "summary": "출결 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "학생 ID",
                        "name": "student_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "날짜 (YYYY-MM-DD)",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Attendance not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Attendance": {
            "description": "출결 정보 모델",
            "type": "object",
            "properties": {
                "check_in": {
                    "type": "string",
                    "example": "13:55"
                },
                "check_out": {
                    "type": "string",
                    "example": "15:35"
                },
                "date": {
                    "type": "string",
                    "example": "2024-01-15"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "출석",
                        "결석",
                        "지각"
                    ],
                    "example": "출석"
                },
                "student_id": {
                    "type": "integer",
                    "example": 1001
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}